Langchain_Q&A_Chatbot/requirements.txt
langchain
python-dotenv
streamlit
fireworks-ai